Split ETL functions into independent modules for more organised unit tests
