+++ src/pipeline/load.py
@@ -1,0 +1,2 @@
+async def load( data ):
+    pass
