+++ src/pipeline/transform.py
@@ -1,0 +1,2 @@
+async def transform( data ):
+    pass
